loading test

+++ projects/remult-admin/src/lib/ui/Table.svelte
@@ -12,6 +12,8 @@
   import EditableRow from './EditableRow.svelte'
   import Filter from './Filter.svelte'
   import { writable, type Writable } from 'svelte/store'
+  import Loading from './Loading.svelte'
+  import { SLoading } from '../stores/SLoading.js'
 
   export let fields: FieldUIInfo[]
   export let relations: EntityRelationToManyInfo[]
@@ -23,7 +25,7 @@
 
   let filter: Writable<EntityFilter<any>> = writable({})
 
-  let items = []
+  let items: any[] | null = null
   let totalRows = -1
   let unSub: (() => void) | null = null
 
@@ -32,18 +34,19 @@
       unSub()
     }
 
+    const where = { $and: [currentFilter, { ...parentRelation }] }
+
+    $SLoading = [...new Set([repo.metadata.key, ...$SLoading])]
+
     unSub = repo
       .liveQuery({
         ...options,
-        where: {
-          $and: [currentFilter, { ...parentRelation }],
-        },
+        where,
       })
       .subscribe(async (info) => {
+        $SLoading = $SLoading.filter((x) => x !== repo.metadata.key)
         items = info.applyChanges(items)
-        totalRows = await repo.count({
-          $and: [currentFilter, { ...parentRelation }],
-        })
+        totalRows = await repo.count(where)
       })
   }
 
@@ -76,20 +79,24 @@
       nav?.classList.toggle('hide-navigation')
     }}
   >
-    <svg
-      class="hamburger-icon"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      ><rect y="4.68134" width="24" height="2.03049" fill="black" /><rect
-        y="10.9847"
+    {#if $SLoading.length === 0}
+      <svg
+        class="hamburger-icon"
         width="24"
-        height="2.03049"
-        fill="black"
-      /><rect y="17.2882" width="24" height="2.03049" fill="black" /></svg
-    >
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+        ><rect y="4.68134" width="24" height="2.03049" fill="black" /><rect
+          y="10.9847"
+          width="24"
+          height="2.03049"
+          fill="black"
+        /><rect y="17.2882" width="24" height="2.03049" fill="black" /></svg
+      >
+    {:else}
+      <Loading />
+    {/if}
   </button>
 
   <div class="page-bar__title title" style="--color: {color}">
@@ -188,18 +195,20 @@
           }}
         />
       {/if}
-      {#each items as row, i (repo.metadata.idMetadata.getId(row))}
-        <EditableRow
-          rowId={repo.metadata.idMetadata.getId(row)}
-          {row}
-          save={async (item) => {
-            await repo.update(row, item)
-          }}
-          deleteAction={() => repo.delete(row)}
-          columns={fields}
-          {relations}
-        />
-      {/each}
+      {#if items}
+        {#each items as row, i (repo.metadata.idMetadata.getId(row))}
+          <EditableRow
+            rowId={repo.metadata.idMetadata.getId(row)}
+            {row}
+            save={async (item) => {
+              await repo.update(row, item)
+            }}
+            deleteAction={() => repo.delete(row)}
+            columns={fields}
+            {relations}
+          />
+        {/each}
+      {/if}
     </tbody>
   </table>
 </div>
